Adds support for deduping logs.
Multiple filters can return the same log, so we need to dedupe logs rather than throwing.  This is a breaking change because we previously had documented behavior that duplicate logs would throw, whereas now we just assert that if the block hash and index are the same, then the log is the same.