Fixes bug where an Ethereum node giving bad answers could break things.
This changes the way we announce block/log additions/removals so that we wait until we are sure we can fully synch with the node before doing any announcements.  This allows us to rollback when the Ethereum node gives us something weird in response such as a claim that a parent block doesn't exist, or logs that don't line up with the expected block.

No longer does console.log when an error occurs, instead calls the optional user provided `errorCallback`.  This cleans up the tests, and also allows the user to do something more useful with errors like track them.

Removes deprecated ES5 callback stuff.